+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2022-08-04
+> **Please see the UPGRADE.md guide for instructions**
+
+### Added
+- Added new helper method `Category()->getByName()`
+- Added new helper method `Media()->upload()`
+- Added examples in the repository. See the `_examples` folder
+
+### Changed
+- The helper methods in the `FixtureHelper` service are now split into multiple smaller units
+  - `Media` now holds all media-related helpers
+  - `Category` now holds all category-related helpers
+  - `SalesChannel` now holds all sales-channel-related helpers
+  - `Customer` now holds all customer-related helpers
+  - `Cms` now holds all CMS-page-related helpers
+  - `PaymentMethod` now holds all payment-method-related helpers
+  - `ShippingMethod` now holds all shipping-method-related helpers
+- This is the first licensed version. See LICENSE file
+
 ## [1.8.0] - 2022-06-03
 ### Added
 - Added FixtureTrait
@@ -85,11 +104,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `getNotSpecifiedSalutationId`
   - `getGermanCountryId`
 
-[1.8.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.7.0...1.8.0
-[1.7.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.6.0...1.7.0
-[1.6.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.5.0...1.6.0
-[1.5.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.4.0...1.5.0
-[1.4.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.3.0...1.4.0
-[1.3.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.2.0...1.3.0
-[1.2.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.1.0...1.2.0
-[1.1.0]: https://gitlab.com/basecom-gmbh/shopware/v6/plugins/FixturePlugin/-/compare/1.0.0...1.1.0
+[2.0.0]: https://github.com/basecom/FixturesPlugin/compare/1.8.0...2.0.0
+[1.8.0]: https://github.com/basecom/FixturesPlugin/compare/1.7.0...1.8.0
+[1.7.0]: https://github.com/basecom/FixturesPlugin/compare/1.6.0...1.7.0
+[1.6.0]: https://github.com/basecom/FixturesPlugin/compare/1.5.0...1.6.0
+[1.5.0]: https://github.com/basecom/FixturesPlugin/compare/1.4.0...1.5.0
+[1.4.0]: https://github.com/basecom/FixturesPlugin/compare/1.3.0...1.4.0
+[1.3.0]: https://github.com/basecom/FixturesPlugin/compare/1.2.0...1.3.0
+[1.2.0]: https://github.com/basecom/FixturesPlugin/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/basecom/FixturesPlugin/compare/1.0.0...1.1.0